requirements.txt
fastapi
uvicorn
openai
supabase
python-dotenv
streamlit
requests
beautifulsoup4
tweepy
pyOpenSSL
